# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DraftCrane API
  description: |
    Backend API for DraftCrane (dc) -- a browser-based nonfiction book writing
    environment with AI assistance. Powered by Cloudflare Workers and Hono.

    ## Authentication

    Most endpoints require a valid **Clerk JWT** passed as a Bearer token in the
    `Authorization` header. Public endpoints (health check, Clerk webhook, OAuth
    callback) use their own verification mechanisms and are documented accordingly.

    ## Rate Limiting

    Standard endpoints enforce **120 requests per minute** per user (atomic
    time-window counters in D1). AI and export endpoints have tighter limits
    noted in their descriptions.

    ## Error Responses

    All errors follow a consistent shape with `error` (human-readable message),
    `code` (machine-readable enum), and `requestId` (ULID for tracing).
  version: 2.0.0
  contact:
    name: DraftCrane Engineering
  license:
    name: Proprietary
servers:
  - url: https://dc-api.automation-ab6.workers.dev
    description: Production
...
components:
  securitySchemes:
    clerkBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk-issued JWT token
  schemas:
    Error:
      type: object
      required:
        - error
        - code
        - requestId
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          enum:
            - AUTH_REQUIRED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - RATE_LIMITED
            - DRIVE_ERROR
            - DRIVE_NOT_CONNECTED
            - VALIDATION_ERROR
            - LAST_CHAPTER
            - AI_ERROR
            - EXPORT_ERROR
            - INTERNAL_ERROR
          description: Machine-readable error code
        requestId:
          type: string
          description: ULID request ID for tracing
    Project:
      type: object
      required:
        - id
        - userId
        - title
        - description
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: ULID identifier
        userId:
          type: string
        title:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 1000
        status:
          type: string
          enum: [active, archived]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectWithChapters:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          required:
            - chapters
            - totalWordCount
          properties:
            chapters:
              type: array
              items:
                $ref: '#/components/schemas/Chapter'
            totalWordCount:
              type: integer
    ProjectSummary:
      type: object
      required:
        - id
        - title
        - status
        - wordCount
        - chapterCount
        - updatedAt
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
        wordCount:
          type: integer
        chapterCount:
          type: integer
        updatedAt:
          type: string
          format: date-time
    Chapter:
      type: object
      required:
        - id
        - projectId
        - title
        - sortOrder
        - r2Key
        - wordCount
        - version
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: ULID identifier
        projectId:
          type: string
        title:
          type: string
          maxLength: 200
        sortOrder:
          type: integer
        r2Key:
          type: string
          nullable: true
        wordCount:
          type: integer
        version:
          type: integer
          description: Optimistic locking version number
        status:
          type: string
          enum: [draft, review, final]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ImportBackupResult:
      type: object
      required:
        - projectId
        - title
        - chapterCount
      properties:
        projectId:
          type: string
        title:
          type: string
        chapterCount:
          type: integer
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Authentication required
            code: AUTH_REQUIRED
            requestId: 01HXYZ...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not found
            code: NOT_FOUND
            requestId: 01HXYZ...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Title is required
            code: VALIDATION_ERROR
            requestId: 01HXYZ...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Rate limit exceeded
            code: RATE_LIMITED
            requestId: 01HXYZ...
...
security:
  - clerkBearerAuth: []
paths:
...
  /projects:
    post:
      tags: [Projects]
      summary: Create a new project
      description: |
        Creates a new project with a default "Chapter 1". Title is required
        (max 500 characters), description is optional (max 1000 characters).
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  maxLength: 500
                description:
                  type: string
                  maxLength: 1000
      responses:
        '201':
          description: Project created with default chapter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'

    get:
      tags: [Projects]
      summary: List active projects
      description: |
        Returns all active projects for the authenticated user with aggregated
        word counts and chapter counts.
      operationId: listProjects
      responses:
        '200':
          description: List of active projects
          content:
            application/json:
              schema:
                type: object
                required:
                  - projects
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProjectSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /projects/import:
    post:
      tags: [Projects]
      summary: Import project from ZIP backup
      description: |
        Imports a project from a ZIP backup file previously generated by
        the backup endpoint. Max file size: 50MB. Must be a .zip file
        containing a valid DraftCrane backup manifest.
      operationId: importProject
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: ZIP backup file (max 50MB)
      responses:
        '201':
          description: Project imported
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportBackupResult'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /projects/{projectId}:
    get:
      tags: [Projects]
      summary: Get project details
      description: |
        Returns project details including the full chapter list with metadata.
      operationId: getProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project with chapters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWithChapters'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

    patch:
      tags: [Projects]
      summary: Update project
      description: |
        Updates a project's title and/or description. At least one field
        must be provided.
      operationId: updateProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 500
                description:
                  type: string
                  maxLength: 1000
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

    delete:
      tags: [Projects]
      summary: Archive project
      description: |
        Soft-deletes a project by setting status to "archived". Drive files
        are preserved per PRD requirements.
      operationId: deleteProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project archived
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /projects/{projectId}/duplicate:
    post:
      tags: [Projects]
      summary: Duplicate a project
      description: |
        Creates a complete copy of a project including all chapters and their
        content. The duplicated project gets a new ID and title suffix.
      operationId: duplicateProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Project duplicated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /projects/{projectId}/backup:
    get:
      tags: [Projects]
      summary: Download project backup
      description: |
        Generates and streams a ZIP backup of the project containing a
        manifest.json and HTML files for each chapter. Rate limited to
        5 requests per minute.
      operationId: downloadProjectBackup
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ZIP backup file
          content:
            application/zip:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
                example: attachment; filename="My Book - 2026-02-19.zip"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ─── Chapters ─────────────────────────────────────────────────────────
...